--- kjy_prj/src/main/java/kjy_prj/ServletInitializer.java
@@ -1,13 +1,0 @@
-package kjy_prj;
-
-import org.springframework.boot.builder.SpringApplicationBuilder;
-import org.springframework.boot.web.servlet.support.SpringBootServletInitializer;
-
-public class ServletInitializer extends SpringBootServletInitializer {
-
-	@Override
-	protected SpringApplicationBuilder configure(SpringApplicationBuilder application) {
-		return application.sources(KjyPrjApplication.class);
-	}
-
-}
