--- kjy_prj/src/main/java/kjy_prj/admin/login/LoginDomain.java
@@ -1,14 +1,0 @@
-package kjy_prj.admin.login;
-
-import lombok.Getter;
-import lombok.Setter;
-import lombok.ToString;
-
-@Setter
-@Getter
-@ToString
-public class LoginDomain {
-	private String admin_id;
-	private String password;
-
-}
