--- kjy_prj/src/main/java/kjy_prj/admin/login/LoginVO.java
@@ -1,14 +1,0 @@
-package kjy_prj.admin.login;
-
-import lombok.Getter;
-import lombok.Setter;
-import lombok.ToString;
-
-@Getter
-@Setter
-@ToString
-public class LoginVO {
-	private String admin_id;
-	private String password;	
-
-}
