--- kjy_prj/src/main/java/kjy_prj/admin/movie/MovieService.java
@@ -1,27 +1,0 @@
-package kjy_prj.admin.movie;
-
-import org.apache.ibatis.exceptions.PersistenceException;
-import org.springframework.stereotype.Service;
-
-@Service
-public class MovieService {
-	
-	/**
-	 * 영화를 추가하는 일을 하는 메소드
-	 * @param mVO
-	 * @return 영화 추가 여부
-	 */
-//	public boolean addMovie(MovieVO mVO) {
-//		boolean flag = false;
-//		
-//		try {
-//			MovieDAO mDAO = MovieDAO.getInstance();
-//			flag = mDAO.insertMovie(mVO) == 1;
-//		} catch(PersistenceException pe) {
-//			pe.printStackTrace();
-//		}//end catch
-//		
-//		return flag;
-//	}//addMovie
-	
-}//class
