--- kjy_prj/src/main/java/kjy_prj/admin/movie/movieMapper.xml
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE mapper
-  PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
-  "https://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="kjy_prj.admin.movie">
-
-<insert id="insertMovie" parameterType="movieVO">
-
-</insert>
-
-</mapper>
