--- kjy_prj/src/main/java/kjy_prj/member/mypage/AJAX/MyMovieListController.java
@@ -1,12 +1,0 @@
-package kjy_prj.member.mypage.AJAX;
-
-import org.springframework.web.bind.annotation.RestController;
-
-/**
- * @author : user
- * @fileName : MyMovieListController
- * @since : 24. 12. 16.
- */
-@RestController
-public class MyMovieListController {
-} // MyMovieListController 끝
