+++ kjy_prj/src/main/java/kr/co/sist/ServletInitializer.java
@@ -1,0 +1,13 @@
+package kr.co.sist;
+
+import org.springframework.boot.builder.SpringApplicationBuilder;
+import org.springframework.boot.web.servlet.support.SpringBootServletInitializer;
+
+public class ServletInitializer extends SpringBootServletInitializer {
+
+	@Override
+	protected SpringApplicationBuilder configure(SpringApplicationBuilder application) {
+		return application.sources(KjyPrjApplication.class);
+	}
+
+}
