+++ kjy_prj/src/main/java/kr/co/sist/admin/login/LoginDomain.java
@@ -1,0 +1,14 @@
+package kr.co.sist.admin.login;
+
+import lombok.Getter;
+import lombok.Setter;
+import lombok.ToString;
+
+@Setter
+@Getter
+@ToString
+public class LoginDomain {
+	private String admin_id;
+	private String password;
+
+}
