+++ kjy_prj/src/main/java/kr/co/sist/admin/login/LoginVO.java
@@ -1,0 +1,14 @@
+package kr.co.sist.admin.login;
+
+import lombok.Getter;
+import lombok.Setter;
+import lombok.ToString;
+
+@Getter
+@Setter
+@ToString
+public class LoginVO {
+	private String admin_id;
+	private String password;	
+
+}
