+++ kjy_prj/src/main/java/kr/co/sist/admin/movie/GenreVO.java
@@ -1,0 +1,15 @@
+package kr.co.sist.admin.movie;
+
+import lombok.Getter;
+import lombok.Setter;
+import lombok.ToString;
+
+//MOVIE_NUM	G_NAME	
+
+@Getter
+@Setter
+@ToString
+public class GenreVO {
+	private int movie_num;
+	private String g_name;
+}
