+++ kjy_prj/src/main/java/kr/co/sist/admin/movie/movieMapper.xml
@@ -1,0 +1,55 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE mapper
+  PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
+  "https://mybatis.org/dtd/mybatis-3-mapper.dtd">
+<mapper namespace="kjy_prj.admin.movie">
+
+	<!-- 총 게시물의 수 얻기 -->
+	<select id="totalCnt" resultType="int" parameterType="searchVO">
+		select count(movie_num) cnt from movie
+		<where>
+			<if test="keyword neq null and keyword neq ''">
+				<choose>
+					<when test="field eq 0">
+						instr(movie_num ,#{keyword}) != 0
+					</when>
+					<otherwise>
+						instr(title_k ,#{keyword}) != 0
+					</otherwise>
+				</choose>
+			</if>
+		</where>
+	</select>
+
+	<select id="selectMovie" resultType="MovieDomain" parameterType="searchVO">
+		select movie_num, title_k, screening_f
+		from(select LPAD(movie_num,5,'0')movie_num, title_k, screening_f, remove_f, input_date,
+		row_number() over(order by movie_num desc) rnum
+		from movie
+		where remove_f = 'N'
+		<if test="keyword neq null and keyword neq ''">
+			<choose>
+				<when test="field eq 0">
+					and instr(movie_num, #{keyword}) != 0
+				</when>
+				<otherwise>
+					and instr(title_k, #{keyword}) != 0
+				</otherwise>
+			</choose>
+		</if>
+		)where rnum between #{startNum} and #{endNum}
+	</select>
+	
+	<insert id="insertMovie" parameterType="movieVO">
+		insert into
+		movie(movie_num, title_k, title_e, subtitle, release_date, price, 
+		showtime, info, screening_f, remove_f, main_image, audience_rating,
+		rate)
+		
+	</insert>
+	<!-- insert into
+		m_board(num,subject,content,writer,ip,pass)
+		values(
+		seq_board.nextval,#{subject},#{content},#{writer},#{ip},#{pass}) -->
+
+</mapper>
