+++ kjy_prj/src/main/java/kr/co/sist/member/mypage/AJAX/MyMovieListController.java
@@ -1,0 +1,12 @@
+package kr.co.sist.member.mypage.AJAX;
+
+import org.springframework.web.bind.annotation.RestController;
+
+/**
+ * @author : user
+ * @fileName : MyMovieListController
+ * @since : 24. 12. 16.
+ */
+@RestController
+public class MyMovieListController {
+} // MyMovieListController 끝
