+++ kjy_prj/src/main/java/kr/co/sist/member/mypage/information/MyInformationController.java
@@ -1,0 +1,33 @@
+package kr.co.sist.member.mypage.information;
+
+import ch.qos.logback.core.model.Model;
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+
+/**
+ * @author : user
+ * @fileName : MyInformationController
+ * @since : 24. 12. 16.
+ */
+@Controller
+@RequestMapping("/myPage/information/")
+public class MyInformationController {
+
+
+
+    @GetMapping("myInformation")
+    public String myInformation(Model model) {
+        return "member/mypage/information/myInformation";
+    }
+
+    @GetMapping("retired")
+    public String retired(Model model) {
+        return "member/mypage/information/retired";
+    }
+
+    @GetMapping("modifyPass")
+    public String modifyPass(Model model) {
+        return "memeber/mypage/information/modifyPass";
+    }
+} // MyInformationController 끝
