+++ kjy_prj/src/main/resources/application-dev.properties
@@ -7,3 +7,6 @@ server.servlet.session.timeout=2h
 spring.mvc.view.prefix="/WEB-INF/views/"
 spring.mvc.view.suffix=.jsp
 
+#security
+key=sist1234
+salt=314253
