+++ kjy_prj/src/main/resources/application-prod.properties
@@ -2,7 +2,7 @@ spring.application.name=kjy_prj
 
 #서버 설정
 server.port=80
-server.servlet.session.timeout=1m
+server.servlet.session.timeout=2h
 
 #view 설정
 spring.mvc.view.prefix=/WEB-INF/views/
@@ -23,3 +23,7 @@ spring.datasource.password=qwe123
 spring.jpa.database-platform=org.hibernate.dialect.OracleDialect
 spring.jpa.show-sql=true
 spring.jpa.hibernate.ddl-auto=none
+
+#security
+key=sist1234
+salt=314253
